requirements.txt
flask
gunicorn
mysql.connector
python-dotenv
requests